pcap: fix conflict between 'size_t' and 'uint32_t'.
The packet (snapshot) length is represented by an 32-bit unsigned
interger where 'size_t' can be a 64-bit value on some architecture.
Replacing the 'size_t' variable by a 'uint32_t' resolves the bit width
conflict and still allow a 4GB frame to be read.

+++ libdabba/include/libdabba/pcap.h
@@ -111,7 +111,7 @@ int pcap_link_type_get(int arp_type, enum pcap_linktype *pcap_link_type);
 ssize_t pcap_write(const int fd, const uint8_t * const pkt,
 		   const size_t pkt_len, const size_t pkt_snaplen,
 		   const uint64_t tv_sec, const uint64_t tv_usec);
-ssize_t pcap_read(const int fd, uint8_t * pkt, const size_t pkt_len);
+ssize_t pcap_read(const int fd, uint8_t * pkt, const uint32_t pkt_len);
 void pcap_destroy(const int fd, const char *const pcap_path);
 int pcap_create(const char *const pcap_path, const enum pcap_linktype linktype);
 int pcap_open(const char *const pcap_path, int flags);
